Refactor e2e test file structure
