+++ __tests__/end-to-end/admin-tests/e2e-admin-constants.js
@@ -1,0 +1,1 @@
+export const HOST = 'http://localhost:4000';
