+++ __tests__/end-to-end/admin-tests/serverside-rendering.test.js
@@ -5,7 +5,9 @@
  */
 import Nightmare from 'nightmare';
 
-import { testPathServersideRender, SIMPLE_TEST_TIMEOUT, ADMIN_HOST as HOST } from '../utilities';
+import { testPathServersideRender } from '__tests__/end-to-end/e2e-utilities';
+import { SIMPLE_TEST_TIMEOUT } from '__tests__/end-to-end/e2e-constants';
+import { HOST } from './e2e-admin-constants';
 
 jest.setTimeout(SIMPLE_TEST_TIMEOUT);
