+++ __tests__/end-to-end/e2e-constants.js
@@ -1,0 +1,1 @@
+export const SIMPLE_TEST_TIMEOUT = 10 * 1000;
