+++ __tests__/end-to-end/gazelle-tests/e2e-gazelle-constants.js
@@ -1,0 +1,1 @@
+export const HOST = 'http://localhost:3000';
