+++ __tests__/end-to-end/gazelle-tests/serverside-rendering.test.js
@@ -1,4 +1,6 @@
-import { testPathServersideRender, SIMPLE_TEST_TIMEOUT, GAZELLE_HOST as HOST } from '../utilities';
+import { testPathServersideRender } from '__tests__/end-to-end/e2e-utilities';
+import { SIMPLE_TEST_TIMEOUT } from '__tests__/end-to-end/e2e-constants';
+import { HOST } from './e2e-gazelle-constants';
 
 jest.setTimeout(SIMPLE_TEST_TIMEOUT);
