+++ package.json
@@ -20,7 +20,8 @@
     ],
     "moduleDirectories": [
       "node_modules",
-      "./src"
+      "./src",
+      "."
     ]
   },
   "repository": {
