+++ webpack.beta.config.js
@@ -35,6 +35,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -103,6 +104,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -145,6 +147,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
