+++ webpack.config.js
@@ -35,6 +35,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -100,6 +101,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -148,6 +150,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx']
   },
