+++ webpack.production.config.js
@@ -37,6 +37,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -108,6 +109,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
@@ -151,6 +153,7 @@ module.exports = [{
     modulesDirectories: [
       'node_modules',
       './src/',
+      '.',
     ],
     extensions: ['', '.js', '.jsx'],
   },
